Moved `is_mersenne_prime` from `factor_` to `primetest`
The `is_mersenne_prime` is a function to test if a number is a Mersenne prime.
It has been moved to `primetest` because it is more appropriate to have it in
`primetest` rather than `factor_`. Also, the Lucas-Lehmer test,
a Mersenne prime number test algorithm, has been implemented.
