+++ sympy/ntheory/tests/test_primetest.py
@@ -4,7 +4,9 @@ from sympy.ntheory.generate import Sieve, sieve
 from sympy.ntheory.primetest import (mr, _lucas_extrastrong_params, is_lucas_prp, is_square,
                                      is_strong_lucas_prp, is_extra_strong_lucas_prp,
                                      proth_test, isprime, is_euler_pseudoprime,
-                                     is_gaussian_prime, is_fermat_pseudoprime, is_euler_jacobi_pseudoprime)
+                                     is_gaussian_prime, is_fermat_pseudoprime, is_euler_jacobi_pseudoprime,
+                                     MERSENNE_PRIME_EXPONENTS, _lucas_lehmer_primality_test,
+                                     is_mersenne_prime)
 
 from sympy.testing.pytest import slow, raises
 from sympy.core.numbers import I
@@ -122,6 +124,21 @@ def test_proth_test():
             assert proth_test(n) == (n in A080076)
         else:
             raises(ValueError, lambda: proth_test(n))
+
+
+def test_lucas_lehmer_primality_test():
+    for p in sieve.primerange(3, 100):
+        assert _lucas_lehmer_primality_test(p) == (p in MERSENNE_PRIME_EXPONENTS)
+
+
+def test_is_mersenne_prime():
+    assert is_mersenne_prime(-3) is False
+    assert is_mersenne_prime(3) is True
+    assert is_mersenne_prime(10) is False
+    assert is_mersenne_prime(127) is True
+    assert is_mersenne_prime(511) is False
+    assert is_mersenne_prime(131071) is True
+    assert is_mersenne_prime(2147483647) is True
 
 
 def test_isprime():
